move the hypervm logic to another file
